Refactor object processing and camera state
