+++ frigate/api/media.py
@@ -37,7 +37,7 @@ from frigate.const import (
     RECORD_DIR,
 )
 from frigate.models import Event, Previews, Recordings, Regions, ReviewSegment
-from frigate.object_processing import TrackedObjectProcessor
+from frigate.track.object_processing import TrackedObjectProcessor
 from frigate.util.builtin import get_tz_modifiers
 from frigate.util.image import get_image_from_recording
 from frigate.util.path import get_event_thumbnail_bytes
