+++ frigate/comms/event_metadata_updater.py
@@ -10,6 +10,8 @@ logger = logging.getLogger(__name__)
 
 class EventMetadataTypeEnum(str, Enum):
     all = ""
+    manual_event_create = "manual_event_create"
+    manual_event_end = "manual_event_end"
     regenerate_description = "regenerate_description"
     sub_label = "sub_label"
