+++ frigate/review/maintainer.py
@@ -25,7 +25,7 @@ from frigate.const import (
 )
 from frigate.events.external import ManualEventState
 from frigate.models import ReviewSegment
-from frigate.object_processing import TrackedObject
+from frigate.track.object_processing import TrackedObject
 from frigate.review.types import SeverityEnum
 from frigate.util.image import SharedMemoryFrameManager, calculate_16_9_crop
